backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
opencv-python==4.8.1.78
numpy==1.26.4
mediapipe==0.10.9
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
bson==0.5.10
